add shared travel insurance module
